pyproject.toml
[project]
name = "mu2e"
version = "0.2.0"
description = "mu2e docdb and LLM interface"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "beautifulsoup4",
    "httpx",
    "pdfplumber",
    "requests",
    "urllib3",
    "tiktoken",
    "numpy",
    "python-dotenv",
    "openai>=1.0.0",
    "anthropic",
    "chromadb",
    "tqdm",
    "aiohttp",
    "mcp[cli]",
    "slack_sdk",
    "pytz",
    "flask",
    "flask-socketio",
    "python-pptx",
    "python-docx",
    "pandas",
    "pillow",
]

[project.optional-dependencies]
minimal = [
    "beautifulsoup4",
    "httpx",
    "pdfplumber",
    "requests",
    "numpy",
    "python-dotenv"
]
slack = [
    "slack_sdk",
    "pytz"
]
mcp = [
    "mcp[cli]>=2.0.0"
]




[build-system]
requires = [ "hatchling",]
build-backend = "hatchling.build"

[project.scripts]
mu2e-docdb = "mu2e.cli.docdb_cli:main"
mu2e-chat = "mu2e.cli.chat_cli:main"
mu2e-slack = "mu2e.cli.slack_cli:main"
mu2e-web = "mu2e.web.app:main"
mu2e-mcp-server = "mu2e.mcp.docdb.server_fastmcp:main"
